refactor(learning): split runtime scenarios and remove integration leftovers
